add GL hpp & cpp, create & use texture from openCV Mat, create some functions for resize the GL window
