+++ src/Game.cpp
@@ -9,14 +9,16 @@
 #include <opencv2/imgproc.hpp>
 #include <opencv2/opencv.hpp>
 
-#define GLFW_INCLUDE_GLU
-#include <GLFW/glfw3.h>
+//#define GLFW_INCLUDE_GLU
+//#include <GLFW/glfw3.h>
 
 #include <vector>
 #include <iostream>
 #include <stdlib.h>
 #include <stdio.h>
 #include <random>
+
+#include "GL.hpp"
 
 #include "Game.hpp"
 #include "Image.hpp"
@@ -25,177 +27,6 @@ using namespace std;
 using namespace cv;
 
 
-void controls_GL(GLFWwindow* GL_window, int key, int scancode, int action, int mods)
-{
-	if (action == GLFW_PRESS)
-		if (key == GLFW_KEY_ESCAPE)
-			glfwSetWindowShouldClose(GL_window, GL_TRUE);
-}
-
-
-void def_carre(void)
-{
-	glBegin(GL_POLYGON);
-	glVertex2f(-0.5, -0.5);
-	glVertex2f(-0.5, 0.5);
-	glVertex2f(0.5, 0.5);
-	glVertex2f(0.5, -0.5);
-	glEnd();
-
-}
-
-void def_axes(void)
-{
-	//	X axe
-	glBegin(GL_LINES);
-	glColor3f(1.0f, 0.0f, 0.0f);
-	glVertex2f(-1.0f, 0.0f);
-	glVertex2f(1.0f, 0.0f);
-	glEnd();
-
-	//	Y axe
-	glBegin(GL_LINES);
-	glColor3f(0.0f, 1.0f, 0.0f);
-	glVertex2f(0.0f, 1.0f);
-	glVertex2f(0.0f, -1.0f);
-	glEnd();
-
-	//	Z axe
-	glBegin(GL_LINES);
-	glColor3f(0.0f, 0.0f, 1.0f);
-	glVertex3f(0.0f, 0.0f, 1.0f);
-	glVertex3f(0.0f, 0.0f, -1.0f);
-	glEnd();
-}
-
-void def_walls(Mat cloud) {
-	glBegin(GL_POINTS);
-	glColor3f(1.0f, 1.0f, 1.0f);
-
-	for (int row = 0; row < cloud.rows; row++) {
-		for (int col = 0; col < cloud.cols; col++) {
-			uchar intensity = cloud.at<uchar>(row, col);
-			
-			if (intensity != 0) {
-				float x = (float)col / (float)cloud.cols;
-				float y = -(float)row / (float)cloud.rows;
-				//cout << "X : " << x << " Y : " << y << endl;
-				glVertex3f(x, y, 0);
-			}
-
-		}
-	}
-	glEnd();
-
-}
-
-vector<uchar> Mat_to_Array(Mat img) {
-	vector<uchar> vec_img;
-	for (int col = 0; col < img.cols; col++) {
-		for (int row = 0; row < img.rows; row++) {
-
-			Vec3b color = img.at<Vec3b>(row, col);
-
-			vec_img.push_back(color[2]); // R
-			vec_img.push_back(color[1]); // G
-			vec_img.push_back(color[0]); // B
-
-
-		}
-	}
-
-	return vec_img;
-}
-
-
-
-
-GLFWwindow* init_GL(int width = 800, int height = 800) {
-
-	//	Init the library
-	if (!glfwInit()) {
-		cerr << "Initialisation GLFW Failed" << endl;
-		return NULL;
-	}
-
-	//	Create a windowed mode window and its OpenGL context 
-	GLFWwindow* GL_window = glfwCreateWindow(width, height, "GL_window", NULL, NULL);
-	if (!GL_window)
-	{
-		cerr << "Initialisation of GL window Failed" << endl;
-		glfwTerminate();
-		return NULL;
-	}
-
-	/* Make the window's context current */
-	glfwMakeContextCurrent(GL_window);
-
-	glfwSetKeyCallback(GL_window, controls_GL);
-
-	// Get info of GPU and supported OpenGL version
-	cout << "Renderer : " << glGetString(GL_RENDERER) << endl;
-	cout << "OpenGL version supported : " << glGetString(GL_VERSION) << endl;
-
-
-	glEnable(GL_DEPTH_TEST); // Depth Testing
-
-	//glDepthFunc(GL_LEQUAL);
-	//glDisable(GL_CULL_FACE);
-	//glCullFace(GL_BACK);
-
-	return GL_window;
-};
-
-
-
-int draw_GL(GLFWwindow *GL_window, Mat homography, Mat cloud) {
-
-	// Scale to window size
-	GLint windowWidth, windowHeight;
-	glfwGetWindowSize(GL_window, &windowWidth, &windowHeight);
-	glViewport(0, 0, windowWidth, windowHeight);
-
-	glClearColor(0.0f, 0.0f, 0.0f, 0.0f);
-	glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-
-	//	Set PROJECTION
-	glMatrixMode(GL_PROJECTION);
-	glLoadIdentity();
-	gluPerspective(100, (double)windowWidth / (double)windowHeight, 0.1f, 100.0f);
-
-	//	Set MODELVIEW
-	glMatrixMode(GL_MODELVIEW);
-	gluLookAt(0.0, 0.0, 1.0, 0.0, 0.0, 0.0, 0.0, 1.0, 0.0);
-
-
-	while (!glfwWindowShouldClose(GL_window)) {
-
-
-		glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-
-		/*
-		*	DRAW STUFF
-		*/
-		def_axes();
-
-		glPushMatrix();
-		glTranslatef(-0.5f, 0.5f, 0.0f);
-		def_walls(cloud);
-		glPopMatrix();
-
-		// Update Screen
-		glfwSwapBuffers(GL_window);
-
-		// Check for any input, or window movement
-		glfwPollEvents();
-
-
-	}
-
-
-	glfwTerminate();
-	return 0;
-}
 
 
 
@@ -209,7 +40,7 @@ void Game::start(int argc, char **argv) {
 	// echequier
 	int blockSize = 40;
 	int imageSize = blockSize * 8;
-	Mat chessBoard(imageSize, imageSize, CV_8UC1, Scalar::all(0));
+	Mat chessBoard(imageSize, imageSize, CV_8UC3, Scalar::all(0));
 	unsigned char color = 0;
 
 	for (int i = 0; i < imageSize; i = i + blockSize) {
